third_party/blink/web_tests/fast/css/webkit-mask-crash-implicit.html

<!DOCTYPE html>
<head>
<script src="../../resources/js-test.js"></script>
<style>
.box {
mask-repeat: repeat-x;
}
</style>
<script>
if (window.internals)
    testRunner.dumpAsText();
shouldBeEqualToString("document.styleSheets[1].rules[0].style.cssText", "mask-repeat: repeat-x;");
shouldBeEqualToString("document.styleSheets[1].rules[0].style.getPropertyValue('-webkit-mask')", "");
shouldBeEqualToString("document.styleSheets[1].rules[0].style.getPropertyValue('mask-repeat')", "repeat-x");
</script>
</head>
<body>
</body>
</html>
